requirements.txt
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
colorama==0.4.6
tabulate==0.9.0
faker==20.1.0 